+++ scraping-kit-api/src/main/scala/ru/fediq/scrapingkit/model/PageRef.scala
@@ -1,9 +1,10 @@
 package ru.fediq.scrapingkit.model
 
-import akka.http.scaladsl.model.Uri
+import akka.http.scaladsl.model.{HttpMethod, HttpMethods, Uri}
 
 case class PageRef(
   uri: Uri,
+  method: HttpMethod,
   scraperName: String,
   depth: Int = 1,
   context: Map[String, String] = Map.empty,
@@ -14,6 +15,7 @@ case class PageRef(
   def redirectSteps = if (redirectsChain.isEmpty) Nil else uri :: redirectsChain.tail
 
   def chain(nextLocation: Uri) = copy(
+    method = HttpMethods.GET,
     redirectsChain = nextLocation :: redirectsChain
   )
 }