+++ scraping-kit-plugin-rmq/src/main/scala/ru/fediq/scrapingkit/backend/RmqFifoLinksQueue.scala
@@ -54,7 +54,7 @@ class RmqFifoLinksQueue(
   }
 
   override def enqueue(ref: PageRef) = {
-    val body = SerializedPageRef(ref.lastUri.toString(), ref.scraperName, ref.depth, ref.context)
+    val body = SerializedPageRef(ref.lastUri.toString(), ref.method.value, ref.scraperName, ref.depth, ref.context)
       .toJson
       .compactPrint
       .getBytes(StandardCharsets.UTF_8)